packages/langchain-llmos/pyproject.toml
[tool.poetry]
name = "langchain-llmos"
version = "0.1.0"
description = "LangChain integration for LLMOS Bridge — auto-generate tools from the Capability Manifest"
authors = ["LLMOS Bridge Contributors"]
license = "Apache-2.0"
readme = "README.md"
homepage = "https://llmos-bridge.io"
repository = "https://github.com/llmos-bridge/llmos-bridge"
keywords = ["langchain", "llm", "agent", "llmos", "iml"]
packages = [{ include = "langchain_llmos" }]

[tool.poetry.dependencies]
python            = "^3.11"
langchain-core    = "^0.2"
httpx             = "^0.27"
pydantic          = "^2.7"
anthropic                = { version = ">=0.25", optional = true }
openai                   = { version = ">=1.0", optional = true }
google-generativeai      = { version = ">=0.7", optional = true }
pyyaml                   = { version = "^6.0", optional = true }

[tool.poetry.extras]
anthropic = ["anthropic"]
openai    = ["openai"]
gemini    = ["google-generativeai"]
yaml      = ["pyyaml"]
all       = ["anthropic", "openai", "google-generativeai", "pyyaml"]

[tool.poetry.group.dev.dependencies]
pytest            = "^8.2"
pytest-asyncio    = "^0.23"
langchain         = "^0.2"
langchain-openai  = "^0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
